requirements.txt
imagecodecs==2023.1.23
numpy==1.23.5
opencv-python-headless==4.7.0.68
openslide-python==1.2.0
Pillow==9.4.0
scipy==1.10.0
spams==2.6.5.4
tifffile==2023.2.2
